Project-001-Hemoglobin-GC-Content-Analysis/requirements.txt
biopython
pandas
matplotlib
seaborn
